refactor: move createScheduleEmbed and createTableScheduleEmbed to separate files in utils folder
